consolidate Node definition
