+++ PathPlanning/TimeBasedPathPlanning/SafeInterval.py
@@ -17,6 +17,7 @@ from PathPlanning.TimeBasedPathPlanning.GridWithDynamicObstacles import (
     Position,
     empty_2d_array_of_lists,
 )
+from PathPlanning.TimeBasedPathPlanning.Node import Node, NodePath
 import heapq
 import random
 from dataclasses import dataclass
@@ -29,69 +30,13 @@ import time
 #     index and interval member variables are just used to track the path found by the algorithm,
 #     and has no effect on the quality of a node.
 @total_ordering
-class Node:
-    position: Position
-    time: int
-    heuristic: int
-    parent_index: int
+class SIPPNode(Node):
     interval: Interval
-
-    """
-    This is what is used to drive node expansion. The node with the lowest value is expanded next.
-    This comparison prioritizes the node with the lowest cost-to-come (self.time) + cost-to-go (self.heuristic)
-    """
-    def __lt__(self, other: object):
-        if not isinstance(other, Node):
-            return NotImplementedError(f"Cannot compare Node with object of type: {type(other)}")
-        return (self.time + self.heuristic) < (other.time + other.heuristic)
-
-    """
-    Equality only cares about position and time. Heuristic and interval will always be the same for a given
-    (position, time) pairing, so they are not considered in equality.
-    """
-    def __eq__(self, other: object):
-        if not isinstance(other, Node):
-            return NotImplemented
-        return self.position == other.position and self.time == other.time
 
 @dataclass
 class EntryTimeAndInterval:
     entry_time: int
     interval: Interval
-
-class NodePath:
-    path: list[Node]
-    positions_at_time: dict[int, Position] = {}
-
-    def __init__(self, path: list[Node]):
-        self.path = path
-        for (i, node) in enumerate(path):
-            if i > 0:
-                # account for waiting in interval at previous node
-                prev_node = path[i-1]
-                for t in range(prev_node.time, node.time):
-                    self.positions_at_time[t] = prev_node.position
-
-            self.positions_at_time[node.time] = node.position
-
-    """
-    Get the position of the path at a given time
-    """
-    def get_position(self, time: int) -> Position | None:
-        return self.positions_at_time.get(time)
-
-    """
-    Time stamp of the last node in the path
-    """
-    def goal_reached_time(self) -> int:
-        return self.path[-1].time
-
-    def __repr__(self):
-        repr_string = ""
-        for i, node in enumerate(self.path):
-            repr_string += f"{i}: {node}\n"
-        return repr_string
-
 
 class SafeIntervalPathPlanner:
     grid: Grid
@@ -117,16 +62,16 @@ class SafeIntervalPathPlanner:
 
         safe_intervals = self.grid.get_safe_intervals()
 
-        open_set: list[Node] = []
+        open_set: list[SIPPNode] = []
         first_node_interval = safe_intervals[self.start.x, self.start.y][0]
         heapq.heappush(
-            open_set, Node(self.start, 0, self.calculate_heuristic(self.start), -1, first_node_interval)
+            open_set, SIPPNode(self.start, 0, self.calculate_heuristic(self.start), -1, first_node_interval)
         )
 
-        expanded_list: list[Node] = []
+        expanded_list: list[SIPPNode] = []
         visited_intervals = empty_2d_array_of_lists(self.grid.grid_size[0], self.grid.grid_size[1])
         while open_set:
-            expanded_node: Node = heapq.heappop(open_set)
+            expanded_node: SIPPNode = heapq.heappop(open_set)
             if verbose:
                 print("Expanded node:", expanded_node)
 
@@ -138,7 +83,7 @@ class SafeIntervalPathPlanner:
             if expanded_node.position == self.goal:
                 print(f"Found path to goal after {len(expanded_list)} expansions")
                 path = []
-                path_walker: Node = expanded_node
+                path_walker: SIPPNode = expanded_node
                 while True:
                     path.append(path_walker)
                     if path_walker.parent_index == -1:
@@ -163,8 +108,8 @@ class SafeIntervalPathPlanner:
     Generate list of possible successors of the provided `parent_node` that are worth expanding
     """
     def generate_successors(
-        self, parent_node: Node, parent_node_idx: int, intervals: np.ndarray, visited_intervals: np.ndarray
-    ) -> list[Node]:
+        self, parent_node: SIPPNode, parent_node_idx: int, intervals: np.ndarray, visited_intervals: np.ndarray
+    ) -> list[SIPPNode]:
         new_nodes = []
         diffs = [
             Position(0, 0),
@@ -204,7 +149,7 @@ class SafeIntervalPathPlanner:
                 # new interval can be entered
                 for possible_t in range(max(parent_node.time + 1, interval.start_time), min(current_interval.end_time, interval.end_time)):
                     if self.grid.valid_position(new_pos, possible_t):
-                        new_nodes.append(Node(
+                        new_nodes.append(SIPPNode(
                             new_pos,
                             # entry is max of interval start and parent node time + 1 (get there as soon as possible)
                             max(interval.start_time, parent_node.time + 1),
@@ -229,7 +174,7 @@ class SafeIntervalPathPlanner:
 Adds a new entry to the visited intervals array. If the entry is already present, the entry time is updated if the new
 entry time is better. Otherwise, the entry is added to `visited_intervals` at the position of `expanded_node`.
 """
-def add_entry_to_visited_intervals_array(entry_time_and_interval: EntryTimeAndInterval, visited_intervals: np.ndarray, expanded_node: Node):
+def add_entry_to_visited_intervals_array(entry_time_and_interval: EntryTimeAndInterval, visited_intervals: np.ndarray, expanded_node: SIPPNode):
     # if entry is present, update entry time if better
     for existing_entry_and_interval in visited_intervals[expanded_node.position.x, expanded_node.position.y]:
         if existing_entry_and_interval.interval == entry_time_and_interval.interval:
